Update `Framework` implementation
