+++ packages/@remirror/dom/src/index.ts
@@ -1,2 +1,2 @@
-export type { DomFrameworkProps } from './dom-framework';
+export type { DomFrameworkProps, DomFrameworkOutput } from './dom-framework';
 export { createDomEditor, createDomManager } from './dom';
