+++ packages/@remirror/react/src/react-contexts.ts
@@ -2,12 +2,12 @@ import { createContext } from 'react';
 
 import { i18n } from '@remirror/i18n';
 
-import type { I18nContextProps, RemirrorContextProps } from './react-types';
+import type { I18nContextProps, ReactFrameworkOutput } from './react-types';
 
 /**
  * Creates a ReactContext for the Remirror editor
  */
-export const RemirrorContext = createContext<RemirrorContextProps<any> | null>(null);
+export const RemirrorContext = createContext<ReactFrameworkOutput<any> | null>(null);
 
 /**
  * Create the context for the i18n framework used within remirror.
